+++ src/manim_data_structures/__init__.py
@@ -1,8 +1,11 @@
 __version__ = "0.1.7"
 
-from .m_array import *
 from .m_enum import *
+from .m_array import *
+from .m_element import *
+from .m_array_pointer import *
 from .m_variable import *
+from .m_sliding_window import *
 
 __all__ = [
     "MArrayElement",
