+++ src/manim_data_structures/m_array.py
@@ -6,523 +6,8 @@ from copy import deepcopy
 import numpy as np
 from manim import *
 
+from .m_element import MArrayElement
 from .m_enum import MArrayDirection, MArrayElementComp
-
-class MArrayElement(VGroup):
-    """A class that represents an array element.
-
-    Parameters
-    ----------
-    scene
-        Specifies the scene where the object is to be rendered.
-    mob_square_args
-        Arguments for :class:`~manim.mobject.geometry.polygram.Square` that represents the element body.
-    mob_value_args
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element value.
-    mob_index_args
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element index.
-    mob_label_args
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element label.
-    index_pos
-        Specifies the position of :attr:`__mob_index` w.r.t :attr:`__mob_square`
-    index_gap
-        Specifies the distance between :attr:`__mob_index` and :attr:`__mob_square`.
-    label_pos
-        Specifies the position of :attr:`__mob_label` w.r.t :attr:`__mob_square`.
-    label_gap
-        Specifies the distance between :attr:`__mob_label` and :attr:`__mob_square`.
-    next_to_mob
-        Specifies the placement for :attr:`__mob_square` w.r.t another :class:`MArrayElement`.
-    next_to_dir
-        Specifies the direction of placement for :attr:`__mob_square` w.r.t another :class:`MArrayElement`.
-
-    Attributes
-    ----------
-    __scene : :class:`~manim.scene.scene.Scene`
-        The scene where the object is to be rendered.
-    __mob_square_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.geometry.polygram.Square` that represents the element body.
-    __mob_value_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element value.
-    __mob_index_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element index.
-    __mob_label_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element label.
-    __index_pos : :class:`np.ndarray`
-        The position of :attr:`__mob_index` w.r.t :attr:`__mob_square`
-    __index_gap : :class:`float`
-        The distance between :attr:`__mob_index` and :attr:`__mob_square`.
-    __label_pos : :class:`np.ndarray`
-        The position of :attr:`__mob_label` w.r.t :attr:`__mob_square`.
-    __label_gap : :class:`float`
-        The distance between :attr:`__mob_label` and :attr:`__mob_square`.
-    __mob_square : :class:`~manim.mobject.geometry.polygram.Square`
-        Represents the body of the element.
-    __mob_value : :class:`~manim.mobject.text.text_mobject.Text`
-        Represents the value of the element.
-    __mob_index : :class:`~manim.mobject.text.text_mobject.Text`
-        Represents the index of the element.
-    __mob_label : :class:`~manim.mobject.text.text_mobject.Text`
-        Represents the label of the element.
-    """
-
-    def __init_props(
-        self,
-        scene: Scene,
-        index_pos: np.ndarray,
-        index_gap: float,
-        label_pos: np.ndarray,
-        label_gap: float,
-    ) -> None:
-        """Initializes the attributes for the class.
-
-        Parameters
-        ----------
-        scene
-            Specifies the scene where the object is to be rendered.
-        index_pos
-            Specifies the position of :attr:`__mob_index` w.r.t :attr:`__mob_square`
-        index_gap
-            Specifies the distance between :attr:`__mob_index` and :attr:`__mob_square`.
-        label_pos
-            Specifies the position of :attr:`__mob_label` w.r.t :attr:`__mob_square`.
-        label_gap
-            Specifies the distance between :attr:`__mob_label` and :attr:`__mob_square`.
-        """
-
-        self.__mob_square_props: dict = {
-            "color": BLUE_B,
-            "fill_color": BLUE_D,
-            "fill_opacity": 1,
-            "side_length": 1,
-        }
-        self.__mob_value_props: dict = {"text": "", "color": WHITE, "weight": BOLD}
-        self.__mob_index_props: dict = {"text": "", "color": BLUE_D, "font_size": 32}
-        self.__mob_label_props: dict = {"text": "", "color": BLUE_A, "font_size": 38}
-        self.__scene: Scene = scene
-        self.__index_pos: np.ndarray = index_pos
-        self.__index_gap: float = index_gap
-        self.__label_pos: np.ndarray = label_pos
-        self.__label_gap: float = label_gap
-
-    def __update_props(
-        self,
-        mob_square_args: dict = {},
-        mob_value_args: dict = {},
-        mob_index_args: dict = {},
-        mob_label_args: dict = {},
-    ) -> None:
-        """Updates the attributes of the class.
-
-        Parameters
-        ----------
-        mob_square_args
-            Arguments for :class:`~manim.mobject.geometry.polygram.Square` that represents the element body.
-        mob_value_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element value.
-        mob_index_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element index.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element label.
-        """
-
-        self.__mob_square_props.update(mob_square_args)
-        self.__mob_value_props.update(mob_value_args)
-        self.__mob_index_props.update(mob_index_args)
-        self.__mob_label_props.update(mob_label_args)
-
-        if type(self.__mob_value_props["text"]) != str:
-            self.__mob_value_props["text"] = str(self.__mob_value_props["text"])
-
-        if type(self.__mob_index_props["text"]) != str:
-            self.__mob_index_props["text"] = str(self.__mob_index_props["text"])
-
-        if type(self.__mob_label_props["text"]) != str:
-            self.__mob_label_props["text"] = str(self.__mob_label_props["text"])
-
-    def __init_mobs(
-        self,
-        init_square: bool = False,
-        init_value: bool = False,
-        init_index: bool = False,
-        init_label: bool = False,
-        next_to_mob: "MArrayElement" = None,
-        next_to_dir: np.ndarray = RIGHT,
-    ) -> None:
-        """Initializes the mobjects for the class.
-
-        Parameters
-        ----------
-        init_square
-            If `True`, instantiates a :class:`~manim.mobject.geometry.polygram.Square` and assigns it to :attr:`__mob_square`.
-        init_value
-            If `True`, instantiates a :class:`~manim.mobject.text.text_mobject.Text` and assigns it to :attr:`__mob_value`.
-        init_index
-            If `True`, instantiates a :class:`~manim.mobject.text.text_mobject.Text` and assigns it to :attr:`__mob_index`.
-        init_label
-            If `True`, instantiates a :class:`~manim.mobject.text.text_mobject.Text` and assigns it to :attr:`__mob_label`.
-        next_to_mob
-            Specifies placement for :attr:`__mob_square` w.r.t another :class:`MArrayElement`.
-        next_to_dir
-            Specifies direction of placement for :attr:`__mob_square` w.r.t another :class:`MArrayElement`.
-        """
-
-        if init_square:
-            self.__mob_square: Square = Square(**self.__mob_square_props)
-            if next_to_mob is not None:
-                self.__mob_square.next_to(
-                    next_to_mob.fetch_mob_square(), next_to_dir, 0
-                )
-            self.add(self.__mob_square)
-
-        if init_value:
-            self.__mob_value: Text = Text(**self.__mob_value_props)
-            self.__mob_value.next_to(self.__mob_square, np.array([0, 0, 0]), 0)
-            self.add(self.__mob_value)
-
-        if init_index:
-            self.__mob_index: Text = Text(**self.__mob_index_props)
-            self.__mob_index.next_to(
-                self.__mob_square, self.__index_pos, self.__index_gap
-            )
-            self.add(self.__mob_index)
-
-        if init_label:
-            self.__mob_label: Text = Text(**self.__mob_label_props)
-            self.__mob_label.next_to(
-                self.__mob_square, self.__label_pos, self.__label_gap
-            )
-            self.add(self.__mob_label)
-
-    def __deepcopy__(self, memo):
-        """Deepcopy that excludes attributes specified in `exclude_list`."""
-
-        exclude_list = ["_MArrayElement__scene"]
-
-        cls = self.__class__
-        result = cls.__new__(cls)
-        memo[id(self)] = result
-        for k, v in self.__dict__.items():
-            if k not in exclude_list:
-                setattr(result, k, deepcopy(v, memo))
-        return result
-
-    def __init__(
-        self,
-        scene: Scene,
-        mob_square_args: dict = {},
-        mob_value_args: dict = {},
-        mob_index_args: dict = {},
-        mob_label_args: dict = {},
-        index_pos: np.ndarray = UP,
-        index_gap: float = 0.25,
-        label_pos: np.ndarray = LEFT,
-        label_gap: float = 0.5,
-        next_to_mob: "MArrayElement" = None,
-        next_to_dir: np.ndarray = RIGHT,
-        **kwargs
-    ) -> None:
-        """Initializes the class.
-
-        Parameters
-        ----------
-        scene
-            Specifies the scene where the object is to be rendered.
-        mob_square_args
-            Arguments for :class:`~manim.mobject.geometry.polygram.Square` that represents the element body.
-        mob_value_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element value.
-        mob_index_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element index.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element label.
-        index_pos
-            Specifies the position of :attr:`__mob_index` w.r.t :attr:`__mob_square`
-        index_gap
-            Specifies the distance between :attr:`__mob_index` and :attr:`__mob_square`.
-        label_pos
-            Specifies the position of :attr:`__mob_label` w.r.t :attr:`__mob_square`.
-        label_gap
-            Specifies the distance between :attr:`__mob_label` and :attr:`__mob_square`.
-        next_to_mob
-            Specifies the placement for :attr:`__mob_square` w.r.t another :class:`MArrayElement`.
-        next_to_dir
-            Specifies the direction of placement for :attr:`__mob_square` w.r.t another :class:`MArrayElement`.
-        """
-
-        super().__init__(**kwargs)
-
-        # Initialize props
-        self.__init_props(scene, index_pos, index_gap, label_pos, label_gap)
-
-        # Update props
-        self.__update_props(
-            mob_square_args, mob_value_args, mob_index_args, mob_label_args
-        )
-
-        # Initialize mobjects
-        self.__init_mobs(True, True, True, True, next_to_mob, next_to_dir)
-
-    def fetch_mob_square(self) -> Square:
-        """Fetches the square mobject.
-
-        Returns
-        -------
-        :class:`~manim.mobject.geometry.polygram.Square`
-            :attr:`__mob_square`.
-        """
-
-        return self.__mob_square
-
-    def fetch_mob_value(self) -> Text:
-        """Fetches the value mobject.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            :attr:`__mob_value`.
-        """
-
-        return self.__mob_value
-
-    def fetch_mob_index(self) -> Text:
-        """Fetches the index mobject.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            :attr:`__mob_index`.
-        """
-
-        return self.__mob_index
-
-    def fetch_mob_label(self) -> Text:
-        """Fetches the label mobject.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            :attr:`__mob_label`.
-        """
-
-        return self.__mob_label
-
-    def fetch_mob(self, mob_target: MArrayElementComp) -> Mobject:
-        """Fetches the mobject based on the specified enum.
-
-        Parameters
-        ----------
-        mob_target
-            Specifies the :class:`~manim.mobject.mobject.Mobject` to fetch.
-
-        Returns
-        -------
-        :class:`~manim.mobject.mobject.Mobject`
-            Mobject of the class.
-        """
-
-        if mob_target == MArrayElementComp.BODY:
-            return self.fetch_mob_square()
-        elif mob_target == MArrayElementComp.VALUE:
-            return self.fetch_mob_value()
-        elif mob_target == MArrayElementComp.INDEX:
-            return self.fetch_mob_index()
-        elif mob_target == MArrayElementComp.LABEL:
-            return self.fetch_mob_label()
-        else:
-            return self
-
-    def update_mob_value(
-        self,
-        mob_value_args: dict = {},
-        update_anim: Animation = Write,
-        update_anim_args: dict = {},
-        play_anim: bool = True,
-        play_anim_args: dict = {},
-    ) -> Text:
-        """Re-intializes the value mobject.
-
-        Parameters
-        ----------
-        mob_value_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element value.
-        update_anim
-            Animation to be applied to the updated :attr:`__mob_value`.
-        update_anim_args
-            Arguments for update :class:`~manim.animation.animation.Animation`.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            Updated :attr:`__mob_value`.
-        """
-
-        # Update props of mob_value
-        self.__update_props(mob_value_args=mob_value_args)
-
-        # Remove current mob_value
-        self.remove(self.__mob_value)
-
-        # Initialize new mob_value
-        self.__init_mobs(init_value=True)
-
-        # Add new mob_value to group
-        self.add(self.__mob_value)
-
-        # Animate change
-        if play_anim:
-            self.__scene.play(
-                update_anim(self.__mob_value, **update_anim_args), **play_anim_args
-            )
-
-        return self.__mob_value
-
-    def update_mob_index(
-        self,
-        mob_index_args: dict = {},
-        update_anim: Animation = Write,
-        update_anim_args: dict = {},
-        play_anim: bool = True,
-        play_anim_args: dict = {},
-    ) -> Text:
-        """Re-intializes the index mobject.
-
-        Parameters
-        ----------
-        mob_index_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element index.
-        update_anim
-            Animation to be applied to the updated :attr:`__mob_index`.
-        update_anim_args
-            Arguments for update :class:`~manim.animation.animation.Animation`.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            Updated :attr:`__mob_index`.
-        """
-
-        # Update props of mob_index
-        self.__update_props(mob_index_args=mob_index_args)
-
-        # Remove current mob_index
-        self.remove(self.__mob_index)
-
-        # Initialize new mob_index
-        self.__init_mobs(init_index=True)
-
-        # Add new mob_index to group
-        self.add(self.__mob_index)
-
-        # Animate change
-        if play_anim:
-            self.__scene.play(
-                update_anim(self.__mob_index, **update_anim_args), **play_anim_args
-            )
-
-        return self.__mob_index
-
-    def update_mob_label(
-        self,
-        mob_label_args: dict = {},
-        update_anim: Animation = Write,
-        update_anim_args: dict = {},
-        play_anim: bool = True,
-        play_anim_args: dict = {},
-    ) -> Text:
-        """Re-intializes the label mobject.
-
-        Parameters
-        ----------
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the element label.
-        update_anim
-            Animation to be applied to the updated :attr:`__mob_label`.
-        update_anim_args
-            Arguments for update :class:`~manim.animation.animation.Animation`.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            Updated :attr:`__mob_label`.
-        """
-
-        # Update props of mob_label
-        self.__update_props(mob_label_args=mob_label_args)
-
-        # Remove current mob_label
-        self.remove(self.__mob_label)
-
-        # Initialize new mob_label
-        self.__init_mobs(init_label=True)
-
-        # Add new mob_label to group
-        self.add(self.__mob_label)
-
-        # Animate change
-        if play_anim:
-            self.__scene.play(
-                update_anim(self.__mob_label, **update_anim_args), **play_anim_args
-            )
-
-        return self.__mob_label
-
-    def animate_mob_square(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over square mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_square`.
-        """
-
-        return self.__mob_square.animate
-
-    def animate_mob_value(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over value mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_value`.
-        """
-
-        return self.__mob_value.animate
-
-    def animate_mob_index(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over index mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_index`.
-        """
-
-        return self.__mob_index.animate
-
-    def animate_mob_label(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over label mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_label`.
-        """
-
-        return self.__mob_label.animate
-
 
 class MArray(VGroup):
     """A class that represents an array.
@@ -1505,1196 +990,3 @@ class MArray(VGroup):
             update_indices(play_anim_args=play_anim_args)
 
         return (remove_anim, update_indices)
-
-
-class MArrayPointer(VGroup):
-    """A class that represents a pointer.
-
-    Parameters
-    ----------
-    scene
-        Specifies the scene where the object is to be rendered.
-    arr
-        Specifies the array to which the pointer is to be attached.
-    index
-        Specifies the index of the element to which the pointer is to be attached.
-    label
-        Specifies the value of the pointer label.
-    arrow_len
-        Specifies the length of :attr:`__mob_arrow`.
-    arrow_gap
-        Specifies the distance between :attr:`__mob_arrow` and :attr:`__arr`.
-    label_gap
-        Specifies the distance between :attr:`__mob_arrow` and :attr:`__mob_label`.
-    pointer_pos
-        Specifies the position of the pointer w.r.t to :attr:`__arr`.
-    mob_arrow_args
-        Arguments for :class:`~manim.mobject.geometry.line.Arrow` that represents the pointer arrow.
-    mob_label_args
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the pointer label.
-    **kwargs
-        Forwarded to constructor of the parent.
-
-    Attributes
-    ----------
-    __scene : :class:`~manim.scene.scene.Scene`
-        The scene where the object is to be rendered.
-    __arr : :class:`~typing.List`\0[:class:`MArrayElement`]
-        The array to which the pointer is attached to.
-    __index : :class:`int`
-        The index of the element to which the pointer is attached to.
-    __label : :class:`str`
-        The value of the pointer label.
-    __arrow_len : :class:`float`
-        The length of :attr:`__mob_arrow`.
-    __arrow_gap : :class:`float`
-        The distance between :attr:`__mob_arrow` and :attr:`__arr`.
-    __label_gap : :class:`float`
-        The distance between :attr:`__mob_arrow` and :attr:`__mob_label`.
-    __pointer_pos : :class:`.m_enum.MArrayDirection`
-        The position of the pointer w.r.t to :attr:`__arr`.
-    __mob_arrow_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.geometry.line.Arrow` that represents the pointer arrow.
-    __mob_label_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the pointer label.
-    __mob_arrow : :class:`~manim.mobject.geometry.line.Arrow`
-        Represents the arrow of the element.
-    __mob_label : :class:`~manim.mobject.text.text_mobject.Text`
-        Represents the label of the element.
-    __updater_pos : :data:`typing.Callable`\0[[], None]
-        The updater function that keeps the pointer intact with the array.
-    """
-
-    __dir_map = [
-        {"np": UP, "dir": MArrayDirection.UP},
-        {"np": DOWN, "dir": MArrayDirection.DOWN},
-        {"np": RIGHT, "dir": MArrayDirection.RIGHT},
-        {"np": LEFT, "dir": MArrayDirection.LEFT},
-    ]
-    """Maps :class:`~.m_enum.MArrayDirection` to :class:`np.ndarray`."""
-
-    def __calc_arrow_pos(self) -> np.ndarray:
-        """Calculates direction vector for the arrow mobject.
-
-        Returns
-        -------
-        :class:`np.ndarray`
-            Position vector for :attr:`__mob_arrow`.
-        """
-
-        arr_dir_np = self.__dir_map[self.__arr.fetch_arr_dir().value]["np"]
-        arrow_pos_np = np.copy(self.__dir_map[self.__pointer_pos.value]["np"])
-
-        # If array's direction and pointer's direction are not perpendicular to each other
-        if np.dot(arr_dir_np, arrow_pos_np):
-            # swap the x and y values of arrow_pos_np
-            arrow_pos_np[0], arrow_pos_np[1] = arrow_pos_np[1], arrow_pos_np[0]
-            # update the __pointer_pos accordingly
-            self.__pointer_pos = self.__dir_map[
-                (self.__pointer_pos.value + 2) % len(self.__dir_map)
-            ]["dir"]
-
-        return arrow_pos_np
-
-    def __add_updater(self) -> None:
-        """Attaches the position updater function with the pointer."""
-
-        def updater_pos(mob: Mobject) -> None:
-            self.__init_pos()
-
-        self.__updater_pos = updater_pos
-
-        self.add_updater(self.__updater_pos)
-
-    def __remove_updater(self) -> None:
-        """Removes the attached position updater function from the pointer."""
-
-        self.remove_updater(self.__updater_pos)
-
-    def __calc_shift_np(self, new_index: int) -> np.ndarray:
-        """Calculates how much the pointer should shift by to point to the new index.
-
-        Parameters
-        ----------
-        new_index
-            Specifies the prospective index of element to which the pointer is to be attached.
-
-        Returns
-        -------
-        :class:`np.ndarray`
-            A vector that represents how much the pointer should shift.
-        """
-
-        to_lesser_index = False
-        index_start = self.__index
-        index_end = new_index
-        if index_start > index_end:
-            index_start, index_end = index_end, index_start
-            to_lesser_index = True
-
-        return (
-            (
-                self.__arr._MArray__sum_elem_len(index_start, index_end)
-                - (
-                    self.__arr.fetch_mob_arr()[self.__index]
-                    .fetch_mob_square()
-                    .side_length
-                )
-            )
-            * self.__dir_map[self.__arr.fetch_arr_dir().value]["np"]
-            * (-1 if to_lesser_index else 1)
-        )
-
-    def __init_props(
-        self,
-        scene: Scene,
-        arr: MArray,
-        index: int,
-        label: str,
-        arrow_len: float,
-        arrow_gap: float,
-        label_gap: float,
-        pointer_pos: MArrayDirection,
-    ) -> None:
-        """Initializes the attributes for the class.
-
-        Parameters
-        ----------
-        scene
-            Specifies the scene where the object is to be rendered.
-        arr
-            Specifies the array to which the pointer is to be attached.
-        index
-            Specifies the index of the element to which the pointer is to be attached.
-        label
-            Specifies the value of the pointer label.
-        arrow_len
-            Specifies the length of :attr:`__mob_arrow`.
-        arrow_gap
-            Specifies the distance between :attr:`__mob_arrow` and :attr:`__arr`.
-        label_gap
-            Specifies the distance between :attr:`__mob_arrow` and :attr:`__mob_label`.
-        pointer_pos
-            Specifies the position of the pointer w.r.t to :attr:`__arr`.
-        """
-
-        self.__mob_arrow_props: dict = {"color": GOLD_D}
-        self.__mob_label_props: dict = {"text": label, "color": GOLD_A, "font_size": 38}
-        self.__scene: Scene = scene
-        self.__arr: MArray = arr
-        if index >= len(self.__arr.fetch_mob_arr()) or index < 0:
-            raise Exception("Index out of bounds!")
-        self.__index: int = index
-        self.__label: str = label
-        self.__arrow_len: float = arrow_len
-        self.__arrow_gap: float = arrow_gap
-        self.__label_gap: float = label_gap
-        self.__pointer_pos: MArrayDirection = pointer_pos
-
-    def __update_props(
-        self, mob_arrow_args: dict = {}, mob_label_args: dict = {}
-    ) -> None:
-        """Updates the attributes of the class.
-
-        Parameters
-        ----------
-        mob_arrow_args
-            Arguments for :class:`~manim.mobject.geometry.line.Arrow` that represents the pointer arrow.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the pointer label.
-        """
-
-        self.__mob_arrow_props.update(mob_arrow_args)
-        self.__mob_label_props["text"] = self.__label
-        self.__mob_label_props.update(mob_label_args)
-
-        if type(self.__mob_label_props["text"]) != str:
-            self.__mob_label_props["text"] = str(self.__mob_label_props["text"])
-
-    def __init_mobs(self, init_arrow: bool = False, init_label: bool = False) -> None:
-        """Initializes the mobjects for the class.
-
-        Parameters
-        ----------
-        init_arrow
-            If `True`, instantiates a :class:`~manim.mobject.geometry.line.Arrow` and assigns it to :attr:`__mob_arrow`.
-        init_label
-            If `True`, instantiates a :class:`~manim.mobject.text.text_mobject.Text` and assigns it to :attr:`__mob_label`.
-        """
-
-        if init_arrow:
-            arrow_pos_np = self.__calc_arrow_pos()
-            self.__mob_arrow = Arrow(
-                start=(-arrow_pos_np + (arrow_pos_np * self.__arrow_len)),
-                end=-arrow_pos_np,
-                **self.__mob_arrow_props
-            )
-            self.__mob_arrow.next_to(
-                self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square(),
-                arrow_pos_np,
-                self.__arrow_gap,
-            )
-            self.add(self.__mob_arrow)
-
-        if init_label:
-            self.__mob_label = Text(**self.__mob_label_props)
-            self.__mob_label.next_to(
-                self.__mob_arrow,
-                self.__dir_map[self.__pointer_pos.value]["np"],
-                self.__label_gap,
-            )
-            self.add(self.__mob_label)
-
-    def __init_pos(self) -> None:
-        """Initializes the position of the object"""
-
-        arrow_pos_np = self.__calc_arrow_pos()
-        self.next_to(
-            self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square(),
-            arrow_pos_np,
-            self.__arrow_gap,
-        )
-
-    def __deepcopy__(self, memo):
-        """Deepcopy that excludes attributes specified in `exclude_list`."""
-
-        exclude_list = ["_MArrayPointer__scene", "_MArrayPointer__arr"]
-
-        cls = self.__class__
-        result = cls.__new__(cls)
-        memo[id(self)] = result
-        for k, v in self.__dict__.items():
-            if k not in exclude_list:
-                setattr(result, k, deepcopy(v, memo))
-        return result
-
-    def __init__(
-        self,
-        scene: Scene,
-        arr: MArray,
-        index: int = 0,
-        label: str = "",
-        arrow_len: float = 1,
-        arrow_gap: float = 0.25,
-        label_gap: float = 0.25,
-        pointer_pos: MArrayDirection = MArrayDirection.DOWN,
-        mob_arrow_args: dict = {},
-        mob_label_args: dict = {},
-        **kwargs
-    ) -> None:
-        """Initializes the class.
-
-        Parameters
-        ----------
-        scene
-            Specifies the scene where the object is to be rendered.
-        arr
-            Specifies the array to which the pointer is to be attached.
-        index
-            Specifies the index of the element to which the pointer is to be attached.
-        label
-            Specifies the value of the pointer label.
-        arrow_len
-            Specifies the length of :attr:`__mob_arrow`.
-        arrow_gap
-            Specifies the distance between :attr:`__mob_arrow` and :attr:`__arr`.
-        label_gap
-            Specifies the distance between :attr:`__mob_arrow` and :attr:`__mob_label`.
-        pointer_pos
-            Specifies the position of the pointer w.r.t to :attr:`__arr`.
-        mob_arrow_args
-            Arguments for :class:`~manim.mobject.geometry.line.Arrow` that represents the pointer arrow.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the pointer label.
-        **kwargs
-            Forwarded to constructor of the parent.
-        """
-
-        super().__init__(**kwargs)
-
-        # Initialize props
-        self.__init_props(
-            scene, arr, index, label, arrow_len, arrow_gap, label_gap, pointer_pos
-        )
-
-        # Update props
-        self.__update_props(mob_arrow_args, mob_label_args)
-
-        # Initialize mobjects
-        self.__init_mobs(True, True)
-
-        # Add updater
-        self.__add_updater()
-
-    def fetch_mob_arrow(self) -> Arrow:
-        """Fetches the arrow mobject of the pointer.
-
-        Returns
-        -------
-        :class:`~manim.mobject.geometry.line.Arrow`
-            :attr:`__mob_arrow`.
-        """
-
-        return self.__mob_arrow
-
-    def fetch_mob_label(self) -> Text:
-        """Fetches the label mobject of the pointer.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            :attr:`__mob_label`.
-        """
-
-        return self.__mob_label
-
-    def fetch_index(self) -> int:
-        """Fetches the index that the pointer is attached to.
-
-        Returns
-        -------
-        :class:`int`
-            :attr:`__index`.
-        """
-
-        return self.__index
-
-    def update_mob_label(
-        self,
-        label: str,
-        mob_label_args: dict = {},
-        update_anim: Animation = Write,
-        update_anim_args: dict = {},
-        play_anim: bool = True,
-        play_anim_args: dict = {},
-    ) -> Text:
-        """Updates the pointer label.
-
-        Parameters
-        ----------
-        label
-            New value to be assigned to the pointer label.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the pointer label.
-        update_anim
-            Animation to be applied to the updated pointer label.
-        update_anim_args
-            Arguments for update :class:`~manim.animation.animation.Animation`.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            Updated :attr:`__mob_label`.
-        """
-
-        self.__label = label
-
-        # Update props of mob_label
-        self.__update_props(mob_label_args=mob_label_args)
-
-        # Remove current mob_label
-        self.remove(self.__mob_label)
-
-        # Initialize new mob_label
-        self.__init_mobs(init_label=True)
-
-        # Add new mob_label to group
-        self.add(self.__mob_label)
-
-        # Animate change
-        if play_anim:
-            self.__scene.play(
-                update_anim(self.__mob_label, **update_anim_args), **play_anim_args
-            )
-
-        return self.__mob_label
-
-    def animate_mob_arrow(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over arrow mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_arrow`.
-        """
-
-        return self.__mob_arrow.animate
-
-    def animate_mob_label(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over label mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_label`.
-        """
-
-        return self.__mob_label.animate
-
-    def shift_to_elem(
-        self, index: int, play_anim: bool = True, play_anim_args: dict = {}
-    ) -> ApplyMethod:
-        """Shifts pointer to the specified element.
-
-        Parameters
-        ----------
-        index
-            Specifies the index of the element to which the pointer is to be shifted.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.animation.transform.ApplyMethod`
-            Shift animation.
-        """
-
-        if index < 0 or index > len(self.__arr.fetch_mob_arr()):
-            raise Exception("Index out of bounds!")
-
-        shift_anim = ApplyMethod(
-            self.shift, self.__calc_shift_np(index), suspend_mobject_updating=True
-        )
-        self.__index = index
-
-        if play_anim:
-            self.__scene.play(shift_anim, **play_anim_args)
-
-        return shift_anim
-
-    def attach_to_elem(self, index: int) -> None:
-        """Attaches pointer to the specified element.
-
-        Parameters
-        ----------
-        index
-            Specifies the index of the element to which the pointer is to be attached.
-        """
-
-        if index < 0 or index > len(self.__arr.fetch_mob_arr()):
-            raise Exception("Index out of bounds!")
-
-        self.__index = index
-        self.__init_pos()
-
-    def __lt__(self, other: MArrayPointer | int | MVariable) -> bool:
-        """Checks if the pointer is less than the other pointer.
-
-        Parameters
-        ----------
-        other
-            The other pointer.
-
-        Returns
-        -------
-        :class:`bool`
-            `True` if the pointer's pointing index in the `MArray` is less than the other pointer, `False` otherwise.
-        """
-        from .m_variable import MVariable
-        if isinstance(other, MArrayPointer):
-            return self.__index < other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index < other.fetch_value()
-        else:
-            return self.__index < other
-
-    def __le__(self, other: MArrayPointer | int | MVariable) -> bool:
-        """Checks if the pointer is less than or equal to the other pointer.
-
-        Parameters
-        ----------
-        other
-            The other pointer.
-
-        Returns
-        -------
-        :class:`bool`
-            `True` if the pointer's pointing index in the `MArray` is less than or equal to the other pointer, `False` otherwise.
-        """
-        from .m_variable import MVariable
-        if isinstance(other, MArrayPointer):
-            return self.__index <= other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index <= other.fetch_value()
-        else:
-            return self.__index <= other
-
-    def __gt__(self, other: MArrayPointer | int | MVariable) -> bool:
-        """Checks if the pointer is greater than the other pointer.
-
-        Parameters
-        ----------
-        other
-            The other pointer.
-
-        Returns
-        -------
-        :class:`bool`
-            `True` if the pointer's pointing index in the `MArray` is greater than the other pointer, `False` otherwise.
-        """
-        from .m_variable import MVariable
-        if isinstance(other, MArrayPointer):
-            return self.__index > other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index > other.fetch_value()
-        else:
-            return self.__index > other
-
-    def __ge__(self, other: MArrayPointer | int | MVariable) -> bool:
-        """Checks if the pointer is greater than or equal to the other pointer.
-
-        Parameters
-        ----------
-        other
-            The other pointer.
-
-        Returns
-        -------
-        :class:`bool`
-            `True` if the pointer's pointing index in the `MArray` is greater than or equal to the other pointer, `False` otherwise.
-        """
-        from .m_variable import MVariable
-        if isinstance(other, MArrayPointer):
-            return self.__index >= other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index >= other.fetch_value()
-        else:
-            return self.__index >= other
-
-    def __eq__(self, other: MArrayPointer | int | MVariable) -> bool:
-        """Checks if the pointer is equal to the other pointer.
-
-        Parameters
-        ----------
-        other
-            The other pointer.
-
-        Returns
-        -------
-        :class:`bool`
-            `True` if the pointer's pointing index in the `MArray` is equal to the other pointer, `False` otherwise.
-        """
-        from .m_variable import MVariable
-        if isinstance(other, MArrayPointer):
-            return self.__index == other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index == other.fetch_value()
-        else:
-            return self.__index == other
-
-    def __ne__(self, other: MArrayPointer | int | MVariable) -> bool:
-        """Checks if the pointer is not equal to the other pointer.
-
-        Parameters
-        ----------
-        other
-            The other pointer.
-
-        Returns
-        -------
-        :class:`bool`
-            `True` if the pointer's pointing index in the `MArray` is not equal to the other pointer, `False` otherwise.
-        """
-        from .m_variable import MVariable
-        return not self.__eq__(other)
-
-    def __add__(self, other: MArrayPointer | int | MVariable) -> int:
-        """
-        return the addition of the index of the current pointer and the index of the other pointer if other is another pointer.
-        return the addition of the index of the current pointer and the integer if other is an integer.
-        return the addition of the index of the current pointer and the value of the other variable if other is a MVariable.
-
-        Parameters
-        ----------
-        other
-            The value to be added.
-
-        Returns
-        -------
-        :class:`MArrayPointer`
-            The pointer after adding the value.
-        """
-        from .m_variable import MVariable # do this because of circular import
-        if isinstance(other, MArrayPointer):
-            return self.__index + other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index + other.fetch_value()
-        else:
-            return self.__index + other
-    def __sub__(self, other: MArrayPointer | int | MVariable) -> int:
-        """
-        return the subtraction of the index of the current pointer and the index of the other pointer if other is another pointer.
-        return the subtraction of the index of the current pointer and the integer if other is an integer.
-        return the subtraction of the index of the current pointer and the value of the other variable if other is a MVariable.
-        ----------
-        other
-            The value to be subtracted.
-
-        Returns
-        -------
-        :class:`MArrayPointer`
-            The pointer after subtracting the value.
-        """
-        from .m_variable import MVariable
-        if isinstance(other, MArrayPointer):
-            return self.__index - other.fetch_index()
-        elif isinstance(other, MVariable):
-            return self.__index - other.fetch_value()
-        else:
-            return self.__index - other
-
-    def __key(self) -> tuple:
-        # return the tuple of all attributues, used for hashing
-        return (self.__scene, self.__arr, self.__index, self.__label, self.__arrow_len, self.__arrow_gap, self.__label_gap,
-            self.__pointer_pos, self.__updater_pos 
-                )
-    def __hash__(self) -> int:
-        return id(self)
-
-class MArraySlidingWindow(VGroup):
-    """A class that represents a sliding window
-
-    Parameters
-    ----------
-    scene
-        Specifies the scene where the object is to be rendered.
-    arr
-        Specifies the array to which the sliding window is to be attached.
-    index
-        Specifies the index of the element to which the sliding window is to be attached.
-    size
-        Specifies the number of elements the sliding window should enclose.
-    label
-        Specifies the value of the sliding window label.
-    label_gap
-        Specifies the distance between :attr:`__mob_label` and :attr:`__mob_window`.
-    label_pos
-        Specifies the position of the pointer w.r.t to :attr:`__mob_window`.
-    mob_window_args
-        Arguments for :class:`~manim.mobject.geometry.polygram.Rectangle` that represents the window.
-    mob_label_args
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the window label.
-    **kwargs
-        Forwarded to constructor of the parent.
-
-    Attributes
-    ----------
-    __scene : :class:`~manim.scene.scene.Scene`
-        The scene where the object is to be rendered.
-    __arr : :class:`~typing.List`\0[:class:`MArrayElement`]
-        The array to which the sliding window is to be attached.
-    __index : :class:`int`
-        The index of the element to which the sliding window is to be attached.
-    __size : :class:`int`
-        The number of elements the sliding window should enclose.
-    __label : :class:`str`
-        The value of the sliding window label.
-    __label_gap : :class:`float`
-        The distance between :attr:`__mob_label` and :attr:`__mob_window`.
-    __label_pos : :class:`.m_enum.MArrayDirection`
-        The position of the pointer w.r.t to :attr:`__mob_window`.
-    __mob_window_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.geometry.polygram.Rectangle` that represents the window.
-    __mob_label_props : :class:`dict`
-        Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the window label.
-    __mob_window : :class:`~manim.mobject.geometry.polygram.Rectangle`
-        Represents the window of the sliding window.
-    __mob_label : :class:`~manim.mobject.text.text_mobject.Text`
-        Represents the label of the sliding window.
-    __updater_pos : :data:`typing.Callable`\0[[], None]
-        The updater function that keeps the sliding window intact with the array.
-    """
-
-    __dir_map = [
-        {"np": UP, "dir": MArrayDirection.UP},
-        {"np": DOWN, "dir": MArrayDirection.DOWN},
-        {"np": RIGHT, "dir": MArrayDirection.RIGHT},
-        {"np": LEFT, "dir": MArrayDirection.LEFT},
-    ]
-    """Maps :class:`~.m_enum.MArrayDirection` to :class:`np.ndarray`."""
-
-    def __calc_window_dim(self) -> typing.Tuple[float, float]:
-        """Calculates dimensions of window mobject.
-
-        Returns
-        -------
-        :class:`float`
-            Height of :attr:`__mob_window`.
-        :class:`float`
-            Width of :attr:`__mob_window`.
-        """
-
-        height = self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square().side_length
-        width = self.__arr._MArray__sum_elem_len(
-            self.__index, self.__index + self.__size - 1
-        )
-
-        if self.__arr.fetch_arr_dir() in (MArrayDirection.UP, MArrayDirection.DOWN):
-            height, width = width, height
-
-        return (height, width)
-
-    def __calc_window_pos_np(self) -> typing.Tuple[np.ndarray, np.ndarray]:
-        """Calculates position vector and align vector for the window mobject.
-
-        Returns
-        -------
-        :class:`np.ndarray`
-            Position vector for :attr:`__mob_window`
-        :class:`np.ndarray`
-            Align vector for :attr:`__mob_window`
-        """
-
-        point_np = (
-            self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square().get_left()
-        )
-        align_np = LEFT
-
-        arr_dir = self.__arr.fetch_arr_dir()
-        if arr_dir == MArrayDirection.LEFT:
-            point_np = (
-                self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square().get_right()
-            )
-            align_np = RIGHT
-        elif arr_dir == MArrayDirection.UP:
-            point_np = (
-                self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square().get_bottom()
-            )
-            align_np = DOWN
-        elif arr_dir == MArrayDirection.DOWN:
-            point_np = (
-                self.__arr.fetch_mob_arr()[self.__index].fetch_mob_square().get_top()
-            )
-            align_np = UP
-
-        return (point_np, align_np)
-
-    def __calc_label_pos_np(self) -> np.ndarray:
-        """Calculates position vector for the label mobject.
-
-        Returns
-        -------
-        :class:`np.ndarray`
-            Position vector for :attr:`__mob_label`
-        """
-
-        arr_dir = self.__arr.fetch_arr_dir()
-        # Label position is parallel to array growth direction
-        if np.array_equal(
-            self.__dir_map[self.__label_pos.value]["np"],
-            self.__dir_map[arr_dir.value]["np"],
-        ) or np.array_equal(
-            self.__dir_map[self.__label_pos.value]["np"],
-            -self.__dir_map[arr_dir.value]["np"],
-        ):
-            return self.__dir_map[(self.__label_pos.value + 2) % len(self.__dir_map)][
-                "np"
-            ]
-
-        # Label position is perpendicular to array growth direction
-        else:
-            return self.__dir_map[self.__label_pos.value]["np"]
-
-    def __pos_mobs(self, pos_window: bool = False, pos_label: bool = False) -> None:
-        """Positions mobjects of the class.
-
-        Parameters
-        ----------
-        pos_window
-            If `True`, correctly positions :attr:`__mob_window`.
-        pos_label
-            If `True`, correctly positions :attr:`__mob_label`.
-        """
-
-        if pos_window:
-            point_np, align_np = self.__calc_window_pos_np()
-            self.__mob_window.move_to(point_np, align_np)
-
-        if pos_label:
-            self.__mob_label.next_to(
-                self.__mob_window,
-                self.__calc_label_pos_np(),
-                self.__label_gap,
-            )
-
-    def __add_updater(self) -> None:
-        """Attaches the position updater function with the pointer."""
-
-        def updater_pos(mob: Mobject) -> None:
-            self.__init_pos()
-
-        self.__updater_pos = updater_pos
-
-        self.add_updater(self.__updater_pos)
-
-    def __remove_updater(self) -> None:
-        """Removes the attached position updater function from the pointer."""
-
-        self.remove_updater(self.__updater_pos)
-
-    def __init_props(
-        self,
-        scene: Scene,
-        arr: MArray,
-        index: int,
-        size: int,
-        label: str,
-        label_gap: float,
-        label_pos: MArrayDirection,
-    ) -> None:
-        """Initializes the attributes for the class.
-
-        Parameters
-        ----------
-        scene
-            Specifies the scene where the object is to be rendered.
-        arr
-            Specifies the array to which the sliding window is to be attached.
-        index
-            Specifies the index of the element to which the sliding window is to be attached.
-        size
-            Specifies the number of elements the sliding window should enclose.
-        label
-            Specifies the value of the sliding window label.
-        label_gap
-            Specifies the distance between :attr:`__mob_label` and :attr:`__mob_window`.
-        label_pos
-            Specifies the position of the pointer w.r.t to :attr:`__mob_window`.
-        """
-
-        self.__mob_window_props: dict = {"color": RED_D, "stroke_width": 10}
-        self.__mob_label_props: dict = {"text": label, "color": RED_A, "font_size": 38}
-        self.__scene: Scene = scene
-        self.__arr: MArray = arr
-        if index >= len(self.__arr.fetch_mob_arr()) or index < 0:
-            raise Exception("Index out of bounds!")
-        self.__index: int = index
-        if size < 1 or index + size > len(self.__arr.fetch_mob_arr()):
-            raise Exception("Invalid window size!")
-        self.__size: int = size
-        self.__label: str = label
-        self.__label_gap: float = label_gap
-        self.__label_pos: MArrayDirection = label_pos
-
-    def __update_props(
-        self, mob_window_args: dict = {}, mob_label_args: dict = {}
-    ) -> None:
-        """Updates the attributes of the class.
-
-        Parameters
-        ----------
-        mob_window_args
-            Arguments for :class:`~manim.mobject.geometry.polygram.Rectangle` that represents the window.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the window label.
-        """
-
-        self.__mob_window_props.update(mob_window_args)
-        self.__mob_label_props["text"] = self.__label
-        self.__mob_label_props.update(mob_label_args)
-
-        if type(self.__mob_label_props["text"]) != str:
-            self.__mob_label_props["text"] = str(self.__mob_label_props["text"])
-
-    def __init_mobs(self, init_window: bool = False, init_label: bool = False) -> None:
-        """Initializes the mobjects for the class.
-
-        Parameters
-        ----------
-        init_arrow
-            If `True`, instantiates a :class:`~manim.mobject.geometry.polygram.Rectangle` and assigns it to :attr:`__mob_window`.
-        init_label
-            If `True`, instantiates a :class:`~manim.mobject.text.text_mobject.Text` and assigns it to :attr:`__mob_label`.
-        """
-
-        if init_window:
-            height, width = self.__calc_window_dim()
-            self.__mob_window = Rectangle(
-                height=height, width=width, **self.__mob_window_props
-            )
-            self.__pos_mobs(pos_window=True)
-            self.add(self.__mob_window)
-
-        if init_label:
-            self.__mob_label = Text(**self.__mob_label_props)
-            self.__pos_mobs(pos_label=True)
-            self.add(self.__mob_label)
-
-    def __init_pos(self) -> None:
-        """Initializes the position of the object"""
-
-        self.__pos_mobs(True, True)
-
-    def __deepcopy__(self, memo):
-        """Deepcopy that excludes attributes specified in `exclude_list`."""
-
-        exclude_list = ["_MArraySlidingWindow__scene", "_MArraySlidingWindow__arr"]
-
-        cls = self.__class__
-        result = cls.__new__(cls)
-        memo[id(self)] = result
-        for k, v in self.__dict__.items():
-            if k not in exclude_list:
-                setattr(result, k, deepcopy(v, memo))
-        return result
-
-    def __init__(
-        self,
-        scene: Scene,
-        arr: MArray,
-        index: int = 0,
-        size: int = 1,
-        label: str = "",
-        label_gap: float = 0.5,
-        label_pos: MArrayDirection = MArrayDirection.DOWN,
-        mob_window_args: dict = {},
-        mob_label_args: dict = {},
-        **kwargs
-    ) -> None:
-        """Initializes the class.
-
-        Parameters
-        ----------
-        scene
-            Specifies the scene where the object is to be rendered.
-        arr
-            Specifies the array to which the sliding window is to be attached.
-        index
-            Specifies the index of the element to which the sliding window is to be attached.
-        size
-            Specifies the number of elements the sliding window should enclose.
-        label
-            Specifies the value of the sliding window label.
-        label_gap
-            Specifies the distance between :attr:`__mob_label` and :attr:`__mob_window`.
-        label_pos
-            Specifies the position of the pointer w.r.t to :attr:`__mob_window`.
-        mob_window_args
-            Arguments for :class:`~manim.mobject.geometry.polygram.Rectangle` that represents the window.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the window label.
-        **kwargs
-            Forwarded to constructor of the parent.
-        """
-
-        super().__init__(**kwargs)
-
-        # Initialize props
-        self.__init_props(scene, arr, index, size, label, label_gap, label_pos)
-
-        # Update props
-        self.__update_props(mob_window_args, mob_label_args)
-
-        # Initialize mobjects
-        self.__init_mobs(True, True)
-
-        # Add updater
-        self.__add_updater()
-
-    def fetch_mob_window(self) -> Rectangle:
-        """Fetches the window mobject of the sliding window.
-
-        Returns
-        -------
-        :class:`~manim.mobject.geometry.polygram.Rectangle`
-            :attr:`__mob_window`.
-        """
-
-        return self.__mob_window
-
-    def fetch_mob_label(self) -> Text:
-        """Fetches the label mobject of the sliding window.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            :attr:`__mob_label`.
-        """
-
-        return self.__mob_label
-
-    def update_mob_label(
-        self,
-        label: str,
-        mob_label_args: dict = {},
-        update_anim: Animation = Write,
-        update_anim_args: dict = {},
-        play_anim: bool = True,
-        play_anim_args: dict = {},
-    ) -> Text:
-        """Updates the window label.
-
-        Parameters
-        ----------
-        label
-            New value to be assigned to the window label.
-        mob_label_args
-            Arguments for :class:`~manim.mobject.text.text_mobject.Text` that represents the window label.
-        update_anim
-            Animation to be applied to the updated window label.
-        update_anim_args
-            Arguments for update :class:`~manim.animation.animation.Animation`.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.mobject.text.text_mobject.Text`
-            Updated :attr:`__mob_label`.
-        """
-
-        self.__label = label
-
-        # Update props of mob_label
-        self.__update_props(mob_label_args=mob_label_args)
-
-        # Remove current mob_label
-        self.remove(self.__mob_label)
-
-        # Initialize new mob_label
-        self.__init_mobs(init_label=True)
-
-        # Add new mob_label to group
-        self.add(self.__mob_label)
-
-        # Animate change
-        if play_anim:
-            self.__scene.play(
-                update_anim(self.__mob_label, **update_anim_args), **play_anim_args
-            )
-
-        return self.__mob_label
-
-    def animate_mob_window(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over window mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_window`.
-        """
-
-        return self.__mob_window.animate
-
-    def animate_mob_label(self) -> "_AnimationBuilder":  # type: ignore
-        """Invokes the animate property over label mobject.
-
-        Returns
-        -------
-        :class:`_AnimationBuilder`
-            Animate property of :attr:`__mob_label`.
-        """
-
-        return self.__mob_label.animate
-
-    def shift_to_elem(
-        self, index: int, play_anim: bool = True, play_anim_args: dict = {}
-    ) -> ApplyFunction:
-        """Shifts sliding window to the specified element.
-
-        Parameters
-        ----------
-        index
-            Specifies the index of the element to which the sliding window is to be shifted.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.animation.transform.ApplyFunction`
-            Shift animation.
-        """
-
-        if index >= len(self.__arr.fetch_mob_arr()) or index < 0:
-            raise Exception("Index out of bounds!")
-
-        if self.__size < 1 or index + self.__size > len(self.__arr.fetch_mob_arr()):
-            raise Exception("Invalid window size!")
-
-        self.__index = index
-        return self.resize_window(self.__size, play_anim, play_anim_args)
-
-    def attach_to_elem(self, index: int) -> None:
-        """Attaches pointer to the specified element.
-
-        Parameters
-        ----------
-        index
-            Specifies the index of the element to which the sliding window is to be attached.
-        """
-
-        if index >= len(self.__arr.fetch_mob_arr()) or index < 0:
-            raise Exception("Index out of bounds!")
-
-        if self.__size < 1 or index + self.__size > len(self.__arr.fetch_mob_arr()):
-            raise Exception("Invalid window size!")
-
-        self.__index = index
-        self.__init_pos()
-
-    def resize_window(
-        self, size: int, play_anim: bool = True, play_anim_args: dict = {}
-    ) -> ApplyFunction:
-        """Expands or shrinks the window according to the specified size.
-
-        Parameters
-        ----------
-        size
-            Specifies the number of elements the sliding window should enclose.
-        play_anim
-            If `True`, plays the animation(s).
-        play_anim_args
-            Arguments for :py:meth:`Scene.play() <manim.scene.scene.Scene.play>`.
-
-        Returns
-        -------
-        :class:`~manim.animation.transform.ApplyFunction`
-            Resize animation.
-        """
-
-        if size < 1 or self.__index + size > len(self.__arr.fetch_mob_arr()):
-            raise Exception("Invalid window size!")
-
-        self.__size = size
-
-        # Variables for resize_and_shift method
-        arr_dir = self.__arr.fetch_arr_dir()
-        height, width = self.__calc_window_dim()
-        window_pos_np, window_align_np = self.__calc_window_pos_np()
-        label_pos_np = self.__calc_label_pos_np()
-
-        def resize_and_shift(mob: MArraySlidingWindow) -> MArraySlidingWindow:
-            """Resizes and shifts the sliding window
-
-            Returns
-            -------
-            :class:`MArraySlidingWindow`
-                Represents the modified mobject.
-            """
-
-            if arr_dir in (MArrayDirection.UP, MArrayDirection.DOWN):
-                mob.__mob_window.stretch_to_fit_height(height)
-            else:
-                mob.__mob_window.stretch_to_fit_width(width)
-            mob.__mob_window.move_to(window_pos_np, window_align_np)
-            mob.__mob_label.next_to(mob.__mob_window, label_pos_np, mob.__label_gap)
-            return mob
-
-        resize_anim = ApplyFunction(
-            resize_and_shift, self, suspend_mobject_updating=True
-        )
-
-        if play_anim:
-            self.__scene.play(resize_anim, **play_anim_args)
-
-        return resize_anim
